+++ examples/sdk-next-esm/src/clients/index.ts
@@ -1,0 +1,2 @@
+export * from './smoldot';
+export * from './websocket';