+++ packages/common/src/utils/index.ts
@@ -3,6 +3,7 @@ export * as enums from './enums';
 export * as erc20 from './erc20';
 export * as h160 from './h160';
 export * as hex from './hex';
+export * as log from './log';
 export * as xcm from './xcm';
 export * as xc from './xc';