+++ packages/common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @galacticcouncil/common
 
+## 0.3.0
+
+### Minor Changes
+
+[dfe1aea3]: https://github.com/galacticcouncil/sdk/commit/dfe1aea3
+[f4c70f0d]: https://github.com/galacticcouncil/sdk/commit/f4c70f0d
+
+- [[#dfe1aea3][dfe1aea3]] common: add logger
+- [[#f4c70f0d][f4c70f0d]] bump papi to latest (1.23.2)
+
+## 0.2.0
+
+### Minor Changes
+
+- c4783e5: add pow helper
+
 ## 0.1.3
 
 ### Patch Changes