+++ packages/xc-cfg/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @galacticcouncil/xc-cfg
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @galacticcouncil/xc-core@0.6.0
+
+## 0.4.3
+
+### Patch Changes
+
+- fix assethub dwellir rpcs
+
 ## 0.4.2
 
 ### Patch Changes